+++ src/esrs/postgres/projector.rs
@@ -1,26 +1,23 @@
 use async_trait::async_trait;
 use serde::de::DeserializeOwned;
-use sqlx::Transaction;
+use sqlx::PgConnection;
 use uuid::Uuid;
 
 use crate::esrs::store::StoreEvent;
 
 /// Projector trait that takes a Postgres transaction in order to create a read model
 #[async_trait]
-pub trait Projector<Database, Event, Error>
+pub trait Projector<Event, Error>
 where
-    Database: sqlx::Database,
     Event: DeserializeOwned,
 {
     /// This function projects one event in each read model that implements this trait.
     /// The result is meant to catch generic errors.
-    async fn project(
-        &self,
-        event: &StoreEvent<Event>,
-        transaction: &mut Transaction<'_, Database>,
-    ) -> Result<(), Error>;
+    /// TODO: doc for pg connection
+    async fn project(&self, event: &StoreEvent<Event>, connection: &mut PgConnection) -> Result<(), Error>;
 
     /// Delete the read model entry. It is here because of the eventual need of delete an entire
     /// aggregate.
-    async fn delete(&self, aggregate_id: Uuid, transaction: &mut Transaction<'_, Database>) -> Result<(), Error>;
+    /// TODO: doc for implementation; doc for pg connection
+    async fn delete(&self, aggregate_id: Uuid, connection: &mut PgConnection) -> Result<(), Error>;
 }