+++ CHANGELOG.md
@@ -17,7 +17,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `Aggregate` is now pure. API changed so user have to implement `Aggregate` for logic and `AggregateManager` in 
   order to handle persistence layer.
-- `Projector` second parameter changed from `PoolConnection` to `Transaction`.
+- `Projector` second parameter changed from `Transaction` to `PgConnection`.
+- `PgStore` moved to `esrs::store::postgres` module.
+- `Projector` moved to `esrs::store::postgres` module.
+- `Policy` moved to `esrs::store::postgres` module.
 - `Aggregate::apply_event` `payload` parameter changed from reference to value (`Self::Event`).
 - `AggregateManager::event_store` changed to return a reference to it's associated type `EventStore`.
 
@@ -27,6 +30,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `Sqlite` feature and its implementation.
 - `Identifier` trait.
 - `ProjectorEraser` trait.
+- `EraserStore` trait.
+- `ProjectorStore` trait.
 
 ## [0.6.2]